feat: Update bot.py to enable message content intents
This commit updates the `bot.py` file to enable the `message_content` intent in the `discord.Intents` object. By setting `intents.message_content` to `True`, the bot will now have permission to access the content of messages. This change enhances the bot's functionality by allowing it to process and respond to message content.