feat: Update reminder functionality to accept days, hours, and minutes
The `hatirlatici_ekle` command in the `reminder.py` file has been updated to accept the number of days, hours, and minutes for setting a reminder. This allows for more flexibility in setting reminders. The reminder time is calculated based on the current time and the provided duration. The response message now includes the reminder time in UTC format. If an invalid time is provided, an error message is sent. The `check_reminders` task has been modified to use the current time in UTC for checking if a reminder has expired. The `send_dm` function now uses the current time in UTC for displaying the date and time in the reminder message.

Closes #123